Try to put json object/array in a single line
If the json object or array have only scala values as their children,
jpp tries to put the json object/array put in a single line.
Otherwise(json object/array contains a json object or array),
jpp doesn't try to put it in a single line and prettify it normaly
instead.